services/llm-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
beanie==1.23.6
grpcio==1.60.0
grpcio-tools==1.60.0
google-generativeai==0.3.2
openai==1.6.1
langchain==0.0.350
langchain-google-genai==0.0.8
pandas==2.1.4
numpy==1.25.2
prometheus-client==0.19.0
